--- tests/src/lib.rs
@@ -1,4 +1,0 @@
-#![cfg(test)]
-
-mod bindings;
-mod tests;
