+++ tests_tmp/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "ic-tests"
+version = "0.1.0"
+edition = "2021"
+
+[dev-dependencies]
+alloy = { version = "0.12", features = ["full"] }
+candid = "0.10"
+cargo_metadata = "0.19.1"
+ic-cdk = "0.17.1"
+ic-test = { path = "../../ic-test" }
+lazy_static = "1.4.0"
+pocket-ic = "7.0.0"
+serde = "1.0.219"
+serde_bytes = "0.11.17"
+tokio = { version = "1", features = ["full"] }
