+++ tests_tmp/src/lib.rs
@@ -1,0 +1,4 @@
+#![cfg(test)]
+
+mod bindings;
+mod tests;
